+++ bulk/weekly_exports.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
-Version:     2.19
+Version:     2.20
 Date:        11/01/26
 Name:        weekly_exports.php
 Purpose:     Weekly collection exports
@@ -22,7 +22,7 @@ require('../includes/error_handling.php');
 require('../includes/functions.php');
 $msg   = new Message($appConfig);
 $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
-$obj   = new ImportExport($db, $appConfig, $serverEmail);
+$obj   = new ImportExport($db, $appConfig, $gameRules, $serverEmail);
 $historyExporter = new CollectionHistory($db, $appConfig);
 
 $list = '';